--- a/src/JavaTranspiler.js
+++ b/src/JavaTranspiler.js
@@ -10,54 +10,42 @@
 function __transpileToJavaCommon(code) {
   return code
     .replace(/CSS_UNDEFINED/g, 'CSSConstants.UNDEFINED')
+    .replace(/CSS_JUSTIFY_/g, 'CSSJustify.')
+    .replace(/CSS_ALIGN_/g, 'CSSAlign.')
     .replace(/css_flex_direction_t/g, 'CSSFlexDirection')
     .replace(/css_direction_t/g, 'CSSDirection')
-    .replace(/CSS_DIRECTION_/g, 'CSSDirection.')
-    .replace(/CSS_FLEX_DIRECTION_/g, 'CSSFlexDirection.')
     .replace(/css_align_t/g, 'CSSAlign')
-    .replace(/CSS_ALIGN_/g, 'CSSAlign.')
-    .replace(/CSS_WRAP/g, 'CSSWrap.WRAP')
-    .replace(/CSS_POSITION_/g, 'CSSPositionType.')
     .replace(/css_justify_t/g, 'CSSJustify')
-    .replace(/CSS_JUSTIFY_/g, 'CSSJustify.')
     .replace(/css_dim_t/g, 'MeasureOutput')
     .replace(/bool/g, 'boolean')
-    .replace(/^(\s+)([^\s]+)\s+\+=/gm, '$1$2 = $2 +') // Expand +=
-    .replace(/leading\[([^\]]+)\]/g, 'getLeading($1)')
-    .replace(/trailing\[([^\]]+)\]/g, 'getTrailing($1)')
-    .replace(/pos\[([^\]]+)\]/g, 'getPos($1)')
-    .replace(/dim\[([^\]]+)\]/g, 'getDim($1)')
-    .replace(/isUndefined/g, 'CSSConstants.isUndefined')
+    .replace(/style\[dim/g, 'style.dimensions[dim')
+    .replace(/(style|layout)\.width/g, '$1.dimensions[DIMENSION_WIDTH]')
+    .replace(/(style|layout)\.height/g, '$1.dimensions[DIMENSION_HEIGHT]')
+    .replace(/layout\[dim/g, 'layout.dimensions[dim')
+    .replace(/layout\[pos/g, 'layout.position[pos')
+    .replace(/layout\[leading/g, 'layout.position[leading')
+    .replace(/layout\[trailing/g, 'layout.position[trailing')
+    .replace(/\/\*\(c\)!([^*]+)\*\//g, '')
+    .replace(/var\/\*\(java\)!([^*]+)\*\//g, '$1')
     .replace(/\/\*\(java\)!([^*]+)\*\//g, '$1')
-
-    // Since Java doesn't store its attributes in arrays, we need to use setters/getters to access
-    // the appropriate layout/style fields
-    .replace(
-        /(\w+)\.layout\[((?:getLeading|getPos)\([^\)]+\))\]\s+=\s+([^;]+);/gm,
-        'setLayoutPosition($1, $2, $3);')
-    .replace(
-        /(\w+)\.layout\[((?:getTrailing|getPos)\([^\)]+\))\]\s+=\s+([^;]+);/gm,
-        'setLayoutPosition($1, $2, $3);')
-    .replace(
-        /(\w+)\.layout\.direction\s+=\s+([^;]+);/gm,
-        'setLayoutDirection($1, $2);')
-    .replace(/(\w+)\.layout\[((?:getLeading|getPos)\([^\]]+\))\]/g, 'getLayoutPosition($1, $2)')
-    .replace(/(\w+)\.layout\[((?:getTrailing|getPos)\([^\]]+\))\]/g, 'getLayoutPosition($1, $2)')
-    .replace(
-        /(\w+)\.layout\[(getDim\([^\)]+\))\]\s+=\s+([^;]+);/gm,
-        'setLayoutDimension($1, $2, $3);')
-    .replace(/(\w+)\.layout\[(getDim\([^\]]+\))\]/g, 'getLayoutDimension($1, $2)')
-    .replace(/(\w+)\.style\[((?:getLeading|getPos)\([^\]]+\))\]/g, 'getStylePosition($1, $2)')
-    .replace(/(\w+)\.style\[(getDim\([^\]]+\))\]/g, 'getStyleDimension($1, $2)');
 }
 
 function __transpileSingleTestToJava(code) {
   return __transpileToJavaCommon(code)
+    .replace(/CSS_DIRECTION_/g, 'CSSDirection.')
+    .replace(/CSS_FLEX_DIRECTION_/g, 'CSSFlexDirection.')
+    .replace(/CSS_WRAP/g, 'CSSWrap.WRAP')
+    .replace(/CSS_POSITION_/g, 'CSSPositionType.')
     .replace(/new_test_css_node/g, 'new TestCSSNode')
-    .replace( // style.dimensions[CSS_WIDTH] => style.width
+    .replace( // style.position[CSS_TOP] => style.position[CSSLayout.POSITION_TOP]
+        /(style|layout)\.position\[CSS_(LEFT|TOP|RIGHT|BOTTOM)\]/g,
+        function (str, match1, match2) {
+            return match1 + '.position[POSITION_' + match2 + ']';
+        })
+    .replace( // style.dimensions[CSS_WIDTH] => style.dimensions[CSSLayout.DIMENSION_WIDTH]
         /(style|layout)\.dimensions\[CSS_(WIDTH|HEIGHT)\]/g,
         function (str, match1, match2) {
-            return match1 + '.' + match2.toLowerCase();
+            return match1 + '.dimensions[DIMENSION_' + match2 + ']';
         })
     .replace( // style.maxDimensions[CSS_WIDTH] => style.maxWidth
         /(style|layout)\.maxDimensions\[CSS_(WIDTH|HEIGHT)\]/g,
@@ -68,16 +56,6 @@ function __transpileSingleTestToJava(code) {
         /(style|layout)\.minDimensions\[CSS_(WIDTH|HEIGHT)\]/g,
         function (str, match1, match2) {
             return match1 + '.min' + match2.substr(0, 1).toUpperCase() + match2.substr(1).toLowerCase();
-        })
-    .replace( // layout.position[CSS_TOP] => layout.y
-        /layout\.position\[CSS_(TOP|LEFT)\]/g,
-        function (str, match1) {
-            return 'layout.' + (match1 === 'TOP' ? 'top' : 'left');
-        })
-    .replace( // style.position[CSS_TOP] => style.positionTop
-        /style\.(position)\[CSS_(TOP|BOTTOM|LEFT|RIGHT)\]/g,
-        function (str, match1, match2) {
-            return 'style.' + match1 + match2[0] + match2.substring(1).toLowerCase();
         })
     .replace( // style.margin[CSS_TOP] = 12.3 => style.margin[Spacing.TOP].set(12.3)
         /style\.(margin|border|padding)\[CSS_(TOP|BOTTOM|LEFT|RIGHT|START|END)\]\s+=\s+(-?[\.\d]+)/g,