--- a/src/Layout.c
+++ b/src/Layout.c
@@ -1041,7 +1041,10 @@ static void layoutNodeImpl(css_node_t *node, float parentMaxWidth, css_direction
       paddingAndBorderAxisMain
     );
 
-    needsMainTrailingPos = true;
+    if (mainAxis == CSS_FLEX_DIRECTION_ROW_REVERSE ||
+        mainAxis == CSS_FLEX_DIRECTION_COLUMN_REVERSE) {
+      needsMainTrailingPos = true;
+    }
   }
 
   if (!isCrossDimDefined) {
@@ -1053,7 +1056,10 @@ static void layoutNodeImpl(css_node_t *node, float parentMaxWidth, css_direction
       paddingAndBorderAxisCross
     );
 
-    needsCrossTrailingPos = true;
+    if (crossAxis == CSS_FLEX_DIRECTION_ROW_REVERSE ||
+        crossAxis == CSS_FLEX_DIRECTION_COLUMN_REVERSE) {
+      needsCrossTrailingPos = true;
+    }
   }
 
   // <Loop F> Set trailing position if necessary