--- a/src/Layout.c
+++ b/src/Layout.c
@@ -885,18 +885,18 @@ static void layoutNodeImpl(css_node_t *node, float parentMaxWidth, css_direction
         if (isMainDimDefined) {
           setTrailingPosition(node, child, mainAxis);
         }
-      }
 
-      // Now that we placed the element, we need to update the variables
-      // We only need to do that for relative elements. Absolute elements
-      // do not take part in that phase.
-      if (child->style.position_type == CSS_POSITION_RELATIVE) {
-        // The main dimension is the sum of all the elements dimension plus
-        // the spacing.
-        mainDim += betweenMainDim + getDimWithMargin(child, mainAxis);
-        // The cross dimension is the max of the elements dimension since there
-        // can only be one element in that cross dimension.
-        crossDim = fmaxf(crossDim, boundAxis(child, crossAxis, getDimWithMargin(child, crossAxis)));
+        // Now that we placed the element, we need to update the variables
+        // We only need to do that for relative elements. Absolute elements
+        // do not take part in that phase.
+        if (child->style.position_type == CSS_POSITION_RELATIVE) {
+          // The main dimension is the sum of all the elements dimension plus
+          // the spacing.
+          mainDim += betweenMainDim + getDimWithMargin(child, mainAxis);
+          // The cross dimension is the max of the elements dimension since there
+          // can only be one element in that cross dimension.
+          crossDim = fmaxf(crossDim, boundAxis(child, crossAxis, getDimWithMargin(child, crossAxis)));
+        }
       }
     }
 
@@ -1110,6 +1110,7 @@ static void layoutNodeImpl(css_node_t *node, float parentMaxWidth, css_direction
     // the axis are defined (either both left and right or top and bottom).
     for (ii = 0; ii < 2; ii++) {
       axis = (ii != 0) ? CSS_FLEX_DIRECTION_ROW : CSS_FLEX_DIRECTION_COLUMN;
+
       if (!isUndefined(node->layout.dimensions[dim[axis]]) &&
           !isDimDefined(currentAbsoluteChild, axis) &&
           isPosDefined(currentAbsoluteChild, leading[axis]) &&