--- a/src/Layout.js
+++ b/src/Layout.js
@@ -909,7 +909,10 @@ var computeLayout = (function() {
         paddingAndBorderAxisMain
       );
 
-      needsMainTrailingPos = true;
+      if (mainAxis == CSS_FLEX_DIRECTION_ROW_REVERSE ||
+          mainAxis == CSS_FLEX_DIRECTION_COLUMN_REVERSE) {
+        needsMainTrailingPos = true;
+      }
     }
 
     if (!isCrossDimDefined) {
@@ -921,7 +924,10 @@ var computeLayout = (function() {
         paddingAndBorderAxisCross
       );
 
-      needsCrossTrailingPos = true;
+      if (crossAxis == CSS_FLEX_DIRECTION_ROW_REVERSE ||
+          crossAxis == CSS_FLEX_DIRECTION_COLUMN_REVERSE) {
+        needsCrossTrailingPos = true;
+      }
     }
 
     // <Loop F> Set trailing position if necessary