--- a/src/java/src/com/facebook/csslayout/CSSNode.java
+++ b/src/java/src/com/facebook/csslayout/CSSNode.java
@@ -63,6 +63,8 @@ public class CSSNode {
 
   public int lineIndex = 0;
 
+  CSSNode nextAbsoluteChild;
+
   private @Nullable ArrayList<CSSNode> mChildren;
   private @Nullable CSSNode mParent;
   private @Nullable MeasureFunction mMeasureFunction = null;