This branch has a series of changes that makes css-layout around 60% faster in my preliminary benchmarks on Android. 

Some of these commits also benefit the JS and C implementations as they make the layout algorithm more efficient in some setups.

Here's an overview of the changes:
* Avoid a large amount of method invocations in the layout code by moving the generated Java code closer to how the JS and C implementations work i.e. pre-computed arrays with the decisions for main/cross axis. As a bonus, the Java transpiler a now simpler because the code is mostly analogous to the one in C.
* Inline immutable values so avoid method invocations in some of the loops in the layout code.
* Skip/merge some of the loops to avoid unnecessary work in some setups e.g. no need to traverse absolute children if there are none.
* Reduce the search range for flexible children i.e. quicker loops when flexible children are present.

There's definitely more work to be done in terms of adding faster paths for the most common layout setups. For instance, I think we can do a much better job when a node is a simple vertical/horizontal stack of elements. But this is a good start ;-)

In case you're curious, my benchmarks have two types of layouts: a simple vertical stack of 500 children and a more real world "header"-like layout with avatar+title+menu (one flex child surrounded by non-flex ones). Here are the numbers:

* Vertical stack layout
  * Before: ~20ms
  * After: ~7ms
* Header layout
  * Before: 160μs
  * After: 60μs