@vjeux 2015-09-07
This is pretty awesome that you've been able to do all this without changing any test in C and JS :)

cc @kmagiera for reviewing the Java changes