@lucasr 2015-09-08
I've just updated the commits with an extra optimization that only traverses absolute/flexible children instead of traversing all children and checking for their type. This is accomplished by using a private linked list with no extra allocations. It just requires a couple of pointers in each node which is a good enough trade-off here.